[DEV] SEOS-3024: make field 'size' of OS_KeystoreRamFV_DataSubRecord 32 bit large

+++ os_keystore_ram_fv/include/OS_KeystoreRamFV.h
@@ -33,7 +33,7 @@
 
 typedef struct __attribute__((packed))
 {
-    uint16_t    size;
+    uint32_t    size;
     uint8_t     data[];
 }
 OS_KeystoreRamFV_DataSubRecord;
